video_compressor/requirements.txt
ffmpeg-python==0.2.0
python-dotenv==0.21.1
requests==2.27.1